Add cleanups and reorganize post-rebase
